Adicionando verificação dinâmica de versão e opção para forçar download de node.

+++ rn.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 #https://facebook.github.io/react-native/docs/getting-started.html
+site_node="https://nodejs.org"
 tmp="$(mktemp -d)"
 so="$(uname | tr '[:upper:]' '[:lower:]')"
 
@@ -8,21 +9,30 @@ if [ ! "$so" = "linux" ] && [ ! "$so" = "darwin" ]; then
     exit 0
 fi
 
-#if [ -z "$(which npm 2>/dev/null)" ]; then
-version="v8.9.0"
-echo "Baixando node $version e npm."
-cd "${tmp:?}"
-link="https://nodejs.org/dist/$version/node-$version-"$so"-x64.tar.xz"
-curl -L -# "$link" -o node.tar.xz
-tar -Jxf node.tar.xz
-rm -rf "${HOME:?}"/.node4ns "${HOME:?}"/.npm
-mv "${tmp:?}"/node*/ "$HOME"/.node4ns
-rm "${tmp:?}/node.tar.xz"
-echo 'export PATH='$HOME'/.node4ns/bin/:"$PATH"' > $HOME/.noderc
-echo 'source '$HOME'/.noderc' >> $HOME/.bashrc
-source $HOME/.noderc
+echo "Verificando versão mais recente disponível do node."
+a=$(curl -sL "$site_node/en/download/")
 
-#fi
+version="v"$(echo $a | sed -e 's/.*Version: <strong>\([^<]*\)<.*/\1/')
+echo 'Encontrado node '$version' para download.'
+
+
+if [ "$FORCE_NODE" = "1" ] || [ -z "$(which npm 2>/dev/null)" ] || (( $(node --version | sed "s/v\(.\).*/\1/") > ${version:1:1} )); then
+    echo "Baixando ..."
+    cd "${tmp:?}"
+    link="$site_node/dist/$version/node-$version-"$so"-x64.tar.xz"
+    curl -L -# "$link" -o node.tar.xz
+    tar -Jxf node.tar.xz
+    rm -rf "${HOME:?}"/.node_noroot "${HOME:?}"/.npm
+    mv "${tmp:?}"/node*/ "$HOME"/.node_noroot
+    rm "${tmp:?}/node.tar.xz"
+    echo 'export PATH='$HOME'/.node_noroot/bin/:"$PATH"' > $HOME/.noderc
+    echo 'source '$HOME'/.noderc' >> $HOME/.bashrc
+    source $HOME/.noderc
+else
+    echo 'Versão local já está atualizada.'
+fi
+
+echo -en '\n'
 
 if [ "$so" = "darwin" ]; then #configuracao adicional para mac
     touch $HOME/.bash_profile
@@ -33,13 +43,19 @@ fi
 
 if [ -z "$(which react-native 2>/dev/null)" ]; then
     echo 'Instalando React Native.'
-    npm install -g react-native-cli
+    own=$(ls -lah $(which node) | awk '{print $3}')
+    opt=''
+    if [ ! "$own" = "$(whoami)" ]; then
+        echo 'O node está instalado no diretório raiz, por isso é preciso de acesso root apenas para instalar o módulo ou execute nesse terminal: "export FORCE_NODE=1" e rode esse script novamente.'
+        opt='sudo'
+    fi
+    eval "$opt npm install -g react-native-cli"
 fi
 
 if [ "$so" = "linux" ] && [ -z "$ANDROID_HOME" ]; then
     echo "Instalando Android SDK"
     cd "${tmp:?}"
-    git clone https://github.com/wison27/noroot-android && cd noroot-android && bash ./android.sh sdk
+    rm -rf 'noroot-android'; git clone https://github.com/wison27/noroot-android && cd noroot-android && bash ./android.sh sdk
     cd "${tmp:?}"
     rm -rf noroot-android
 fi
